added booking ability

+++ src/main/java/com/example/demo/dao/BookingRepository.java
@@ -9,6 +9,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface BookingRepository extends CrudRepository<Booking, Long> {
-    List<Booking> findBookingsByPlaceId(int placeId);
-    
+    List<Booking> findBookingsByPlaceId(Long placeId);
 }
